chore(cuda): stream migration refactoring (#317)
Closes INT-6464

## CUDA Explicit Stream Migration

Replace implicit per-thread default stream (PTDS) with explicit
`cudaStream_t` management across the entire CUDA stack. Removes the
`--default-stream=per-thread` compiler flag, introduces `CudaStream`,
`StreamGuard`, and `DeviceContext` abstractions, and threads an explicit
stream through every kernel launch, memory operation, and copy path.

**87 files changed, +3750 / -1758**

See
[cuda-stream-migration-design.md](https://github.com/openvm-org/v2-proof-system/blob/test/stream-migration/cuda-stream-migration-design.md)
for full design rationale.

---

## `openvm-stark-backend` (3 files)

### `ProverDevice` trait — `DeviceCtx` associated type

`ProverDevice` gains an associated type and accessor. GPU devices return
`DeviceContext` (CUDA stream + device ID), CPU/Reference devices return
`()`:

```rust
pub trait ProverDevice<PB, TS>: ... {
    type DeviceCtx: Clone + Send + Sync;
    fn device_ctx(&self) -> &Self::DeviceCtx;
}
```

A convenience alias `EngineDeviceCtx<E>` is added to `engine.rs` to
avoid spelling out the full associated type path.

---

## `openvm-cuda-common` (10 files)

### `stream.rs` — New types

- **`CudaStream`** — RAII wrapper around `cudaStream_t`, created with
`cudaStreamNonBlocking`. Non-blocking `Drop` (calls `cudaStreamDestroy`
only, no sync). Per-stream `Mutex<CudaEvent>` for D2H sync via
`to_host_sync()`.
- **`StreamGuard`** — newtype around `Arc<CudaStream>` for ref-counted
stream ownership. `is_same_stream()` via `Arc::ptr_eq`.
`Deref<Target=CudaStream>`.
- **`DeviceContext`** — thin `(device_id, StreamGuard)` bundle. Created
via `DeviceContext::for_device(id)` which binds the CUDA device before
stream creation.

### `memory_manager/` — Stream-aware allocation records

- `AllocRecord { size, stream: StreamGuard }` — tracks which stream each
small allocation belongs to
- `VpmmRecord { size, stream: StreamGuard }` — same for VPMM large
allocations
- Two-stage `d_free`: bookkeeping + `cudaFreeAsync` under lock,
`StreamGuard` dropped outside lock
- VPMM same-stream matching via `Arc::ptr_eq`; cross-stream reuse gated
by event completion

### `copy.rs` / `d_buffer.rs` — `_on(&DeviceContext)` API

All legacy streamless APIs removed. Only `_on(&DeviceContext)` variants
remain: `to_device_on`, `to_host_on`, `fill_zero_on`,
`with_capacity_on`, `d_malloc_on`, `cuda_memcpy_on`. `DeviceBuffer`
gains `#[cfg(feature = "debug-cuda-stream")] alloc_stream` field for
wrong-stream detection.

### Removed

`cudaStreamPerThread`, `CudaStreamId`, `current_stream_id()`,
`current_stream_sync()`, `default_stream_wait()`, `record_on_this()`,
`CudaThreadCleanup`, `gpu_metrics_span()` (PTDS variant),
`CudaStream::new()` (blocking), global `COPY_EVENT` mutex,
`--default-stream=per-thread` flag.

---

## `openvm-cuda-backend` (50+ files)

### Device model (`device.rs`, `engine.rs`)

- `GpuDevice` renamed to `GpuDeviceConfig` (pure config)
- New `GpuDevice { config: GpuDeviceConfig, device_ctx: DeviceContext }`
— stream-owning device handle
- `GpuEngine::new()` calls `GpuDevice::new().expect(...)` to preserve
infallible `StarkEngine::new()`
- `ProverDevice::DeviceCtx = DeviceContext`, `device_ctx()` returns
`&self.device_ctx`

### CUDA launchers (25 `.cu` files)

Every `extern "C"` launcher gains `cudaStream_t stream` as final
parameter. All kernel launches use `<<<grid, block, 0, stream>>>`.
`CHECK_KERNEL()` unchanged. `cudaDeviceSynchronize()` →
`cudaStreamSynchronize(stream)` in `sponge.cu`, `bn254_poseidon2.cu`.

### Rust FFI wrappers (~14 files in `src/cuda/`)

Every `extern "C"` declaration and safe wrapper gains `stream:
cudaStream_t`.

### Proving pipeline

Intermediate structs (`StackedReductionGpu`, `NttImpl`,
`LogupZerocheckGpu`, etc.) store `device_ctx: DeviceContext`. All
allocation/copy calls use `_on` variants. Init paths (`ntt.rs`,
`batch_ntt_small.rs`) create temporary `DeviceContext` with private
streams.

---

## `benchmarks/fields` (5 files)

`.cu` launchers gain `stream` parameter. `lib.rs` FFI wrappers updated.
Tests use `DeviceContext`.

## `scripts/gen_stub.py`

Added `using cudaStream_t = void*;` typedef for the
`gpu_signature_match` CI ABI checker.

---------

+++ crates/cuda-backend/cuda/src/logup_zerocheck/eval_config.cuh
@@ -68,7 +68,7 @@ inline std::pair<dim3, dim3> eval_constraints_launch_params(
 }
 
 template <bool COSET_PARALLEL>
-inline uint32_t temp_sums_buffer_size(
+inline size_t temp_sums_buffer_size(
     uint32_t buffer_size,
     uint32_t skip_domain,
     uint32_t num_x,
@@ -87,11 +87,11 @@ inline uint32_t temp_sums_buffer_size(
         threads_per_block
     );
     // Output layout: [num_blocks][num_cosets * skip_domain]
-    return grid.x * num_cosets * skip_domain;
+    return static_cast<size_t>(grid.x) * num_cosets * skip_domain;
 }
 
 template <bool COSET_PARALLEL>
-inline uint32_t intermediates_buffer_size(
+inline size_t intermediates_buffer_size(
     uint32_t buffer_size,
     uint32_t skip_domain,
     uint32_t num_x,
@@ -113,7 +113,7 @@ inline uint32_t intermediates_buffer_size(
         threads_per_block
     );
     // Layout: [buffer_size][num_threads][num_cosets]
-    return grid.x * block.x * buffer_size * num_cosets;
+    return static_cast<size_t>(grid.x) * block.x * buffer_size * num_cosets;
 }
 
 } // namespace round0_config_impl
